fix attn mask, better bg

+++ examples/hunyuanvideo/hyvideo/inference.py
@@ -465,7 +465,7 @@ class HunyuanVideoSampler(Inference):
             prompt_mask = data['prompt_mask']
             prompt_embeds_2 = data['prompt_embeds_2']
             prompt_embeds = ms.Tensor(prompt_embeds)
-            prompt_mask = ms.Tensor(prompt_mask)
+            prompt_mask = ms.Tensor(prompt_mask, dtype=ms.int32)
             prompt_embeds_2 = ms.Tensor(prompt_embeds_2)
             
             if self.args.cfg_scale > 1.0:    
@@ -473,7 +473,7 @@ class HunyuanVideoSampler(Inference):
                 negative_prompt_mask = data['negative_prompt_mask']
                 negative_prompt_embeds_2 = data['negative_prompt_embeds_2']
                 negative_prompt_embeds = ms.Tensor(negative_prompt_embeds)
-                negative_prompt_mask = ms.Tensor(negative_prompt_mask)
+                negative_prompt_mask = ms.Tensor(negative_prompt_mask, dtype=ms.int32)
                 negative_prompt_embeds_2 = ms.Tensor(negative_prompt_embeds_2)
             else:
                 negative_prompt_embeds = None
@@ -486,6 +486,7 @@ class HunyuanVideoSampler(Inference):
             negative_prompt_embeds = None
             negative_prompt_mask = None
             negative_prompt_embeds_2 = None
+
 
         # ========================================================================
         # Scheduler
